+++ src/main/java/com/atlauncher/data/json/Mod.java
@@ -28,6 +28,8 @@ import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.constants.Constants;
 import com.atlauncher.data.curseforge.CurseForgeFile;
 import com.atlauncher.data.curseforge.CurseForgeProject;
+import com.atlauncher.data.modrinth.ModrinthProject;
+import com.atlauncher.data.modrinth.ModrinthVersion;
 import com.atlauncher.managers.DialogManager;
 import com.atlauncher.managers.LogManager;
 import com.atlauncher.utils.Hashing;
@@ -87,6 +89,8 @@ public class Mod {
     public String description;
     public CurseForgeProject curseForgeProject;
     public CurseForgeFile curseForgeFile;
+    public ModrinthProject modrinthProject;
+    public ModrinthVersion modrinthVersion;
     public boolean ignoreFailures = false;
 
     @SerializedName(value = "curseforge_project_id", alternate = { "curse_id" })